fix like comment bug and have 2 test cases
